address some int/size_t issues (-Wall)

+++ shedskin/lib/builtin/bytes.cpp
@@ -27,7 +27,7 @@ const char *bytes::c_str() const {
     return this->unit.c_str();
 }
 
-const int bytes::size() const {
+const size_t bytes::size() const {
     return this->unit.size();
 }
